+++ lockerbox.cpp
@@ -13,7 +13,7 @@ Lockerbox::Lockerbox() : Machine(ST_MAX_STATES)
 
 }
 
-uint16_t Lockerbox::StartupTest()
+uint8_t Lockerbox::StartupTest()
 {
 	timer.Assign(TIMER_MODBUS_RTU_POLL, 1000, ModbusPoll);
 	return 1;
